# docs/swagger.json
{"openapi":"3.0.0","paths":{"/seat/{id}":{"get":{"operationId":"SeatController_getSeats","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Seats"]}},"/movie/filter-more-genres":{"get":{"operationId":"MovieController_filterMoviesMoreGenres","parameters":[{"name":"genreIds","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Movies"]}},"/movie/filter-one-genre/{id}":{"get":{"operationId":"MovieController_filterMoviesOneGenre","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Movies"]}},"/movie/all":{"get":{"operationId":"MovieController_getAllMovies","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Movies"]}},"/movie/with-genres":{"get":{"operationId":"MovieController_getAllMoviesWithGenres","parameters":[],"responses":{"200":{"description":""}},"tags":["Movies"]}},"/movie/{id}":{"get":{"operationId":"MovieController_getOneMovie","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Movies"]}},"/movie/soft-delete/{id}":{"delete":{"operationId":"MovieController_softDeleteMovie","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Movies"]}},"/movie/restore/{id}":{"post":{"operationId":"MovieController_restoreMovie","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["Movies"]}},"/movie/create":{"post":{"operationId":"MovieController_createMovie","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovieCreateDTO"}}}},"responses":{"201":{"description":""}},"tags":["Movies"]}},"/genre/all":{"get":{"operationId":"GenreController_getAllGenres","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Genres"]}},"/genre/with-movies":{"get":{"operationId":"GenreController_getAllGenresWithMovies","parameters":[],"responses":{"200":{"description":""}},"tags":["Genres"]}},"/genre/available":{"get":{"operationId":"GenreController_getAvailableGenres","parameters":[],"responses":{"200":{"description":""}},"tags":["Genres"]}},"/genre/unavailable":{"get":{"operationId":"GenreController_getUnavailableGenres","parameters":[],"responses":{"200":{"description":""}},"tags":["Genres"]}},"/genre/{id}":{"get":{"operationId":"GenreController_getOneGenre","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Genres"]}},"/genre/create":{"post":{"operationId":"GenreController_createGenre","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenreCreateDTO"}}}},"responses":{"201":{"description":""}},"tags":["Genres"]}},"/genre/enable/{id}":{"post":{"operationId":"GenreController_enableGenre","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["Genres"]}},"/genre/disable/{id}":{"post":{"operationId":"GenreController_disableGenre","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["Genres"]}},"/user/registration":{"post":{"operationId":"UserController_registerUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegistrationDTO"}}}},"responses":{"201":{"description":""}},"tags":["Users"]}},"/user/{id}":{"patch":{"operationId":"UserController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateDTO"}}}},"responses":{"200":{"description":""}},"tags":["Users"]}},"/user/forgotten-password":{"post":{"operationId":"UserController_forgottenPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgottenPasswordDTO"}}}},"responses":{"201":{"description":""}},"tags":["Users"]}},"/user/reset-password/{id}":{"post":{"operationId":"UserController_resetPassword","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDTO"}}}},"responses":{"201":{"description":""}},"tags":["Users"]}},"/user/change-password":{"post":{"operationId":"UserController_changePassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDTO"}}}},"responses":{"201":{"description":""}},"tags":["Users"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/protected":{"get":{"operationId":"AuthController_getAvailableForBothTypes","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/admin":{"get":{"operationId":"AuthController_getAvailableForAdmin","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/customer":{"get":{"operationId":"AuthController_getAvailableForCustomer","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/movie-screening/create":{"post":{"operationId":"MovieScreeningController_createMovieScreening","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovieScreeningCreateDTO"}}}},"responses":{"201":{"description":""}},"tags":["MovieScreenings"]}},"/movie-screening/{id}":{"patch":{"operationId":"MovieScreeningController_editMovieScreening","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovieScreeningEditDTO"}}}},"responses":{"200":{"description":""}},"tags":["MovieScreenings"]}},"/movie-screening/all":{"get":{"operationId":"MovieScreeningController_getMovieScreenings","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["MovieScreenings"]}},"/movie-screening/soft-delete/{id}":{"delete":{"operationId":"MovieScreeningController_softDeleteMovieScreening","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["MovieScreenings"]}},"/movie-screening/restore/{id}":{"post":{"operationId":"MovieScreeningController_restoreMovieScreening","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["MovieScreenings"]}},"/movie-screening/active":{"get":{"operationId":"MovieScreeningController_getActiveMovieScreenings","parameters":[],"responses":{"200":{"description":""}},"tags":["MovieScreenings"]}},"/movie-screening/filter-one-genre/{id}":{"get":{"operationId":"MovieScreeningController_getActiveMovieScreeningsForOneGenre","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["MovieScreenings"]}},"/movie-screening/seat-exists":{"get":{"operationId":"MovieScreeningController_seatExists","parameters":[{"name":"movieScreeningId","required":true,"in":"query","schema":{"type":"number"}},{"name":"seatId","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["MovieScreenings"]}},"/ticket/{action}":{"post":{"operationId":"TicketController_buyOrReserveTickets","parameters":[{"name":"action","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketsBuyingOrReservationDTO"}}}},"responses":{"201":{"description":""}},"tags":["Tickets"]}},"/ticket/cancel/{id}":{"delete":{"operationId":"TicketController_cancelReservation","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Tickets"]}}},"info":{"title":"CLA","description":"","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:3000"}],"components":{"securitySchemes":{"cookie":{"type":"apiKey","in":"cookie","name":"connect.sid"}},"schemas":{"UserRegistrationDTO":{"type":"object","properties":{"fullName":{"type":"string"},"email":{"type":"string","format":"email"},"password":{"type":"string"}},"required":["fullName","email","password"],"additionalProperties":false},"UserUpdateDTO":{"type":"object","properties":{"fullName":{"type":"string"},"email":{"type":"string","format":"email"},"password":{"type":"string"}},"required":[],"additionalProperties":false},"LoginDTO":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string"}},"required":["email","password"],"additionalProperties":false},"MovieScreeningCreateDTO":{"type":"object","properties":{"dateAndTime":{"type":"string","format":"date-time"},"movieId":{"type":"integer"},"auditoriumId":{"type":"integer"}},"required":["dateAndTime","movieId","auditoriumId"],"additionalProperties":false},"MovieScreeningEditDTO":{"type":"object","properties":{"dateAndTime":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["CANCELED","ACTIVE","FINISHED"]},"movieId":{"type":"integer"},"auditoriumId":{"type":"integer"}},"required":[],"additionalProperties":false},"GenreCreateDTO":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"],"additionalProperties":false},"MovieCreateDTO":{"type":"object","properties":{"name":{"type":"string"},"duration":{"type":"integer"},"genreIds":{"type":"array","items":{"type":"integer"},"minItems":1},"price":{"type":"integer"}},"required":["name","duration","genreIds"],"additionalProperties":false},"TicketsBuyingOrReservationDTO":{"type":"object","properties":{"movieScreeningId":{"type":"integer"},"seatIds":{"type":"array","items":{"type":"integer"},"minItems":1}},"required":["movieScreeningId","seatIds"],"additionalProperties":false},"ForgottenPasswordDTO":{"type":"object","properties":{"email":{"type":"string","minLength":1}},"required":["email"],"additionalProperties":false},"ResetPasswordDTO":{"type":"object","properties":{"email":{"type":"string","minLength":1},"newPassword":{"type":"string","minLength":1}},"required":["email","newPassword"],"additionalProperties":false},"ChangePasswordDTO":{"type":"object","properties":{"newPassword":{"type":"string","minLength":1}},"required":["newPassword"],"additionalProperties":false}}}}